new application to build and save graph has been implemented
